modify: 重构项目结构，添加部分测试
