--- __init__.py
@@ -1,8 +1,0 @@
-# -*- coding: UTF-8 -*-
-'''
-@Project ：SpiderToolBox 
-@File    ：__init__.py.py
-@IDE     ：PyCharm 
-@Author  ：JOUSKA.
-@Date    ：2023/12/14 10:50
-'''
