--- spider_tools/other/__init__.py
@@ -1,1 +1,0 @@
-from spider_tools.other.other_tools import *
