--- test.py
@@ -1,3 +1,0 @@
-import spider_tools as tools
-
-tools.open_js()
