+++ tools/download/__init__.py
@@ -1,0 +1,7 @@
+from tools.download.dowloader import *
+
+__all__ = [
+    "download_video",
+    "download_img",
+    "download_character",
+]
