+++ tools/log/__init__.py
@@ -1,0 +1,8 @@
+from tools.log.logger import *
+
+__all__ = [
+    "info",
+    "debug",
+    "warning",
+    "error",
+]
