+++ tools/ocr/__init__.py
@@ -1,0 +1,8 @@
+from tools.ocr.ocrer import *
+
+__all__ = [
+    "ocr_img",
+    "ocr_slide_with_hole",
+    "ocr_slide_with_clean",
+    "ocr_click_choose",
+]
