+++ tools/other/__init__.py
@@ -1,0 +1,5 @@
+from tools.other.other_tools import *
+
+__all__ = [
+    "open_js",
+]
